+++ statuspage/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class StatusPageConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'statuspage'
+    verbose_name = 'Status Pages'